Create an second, small, partition in the img file that can be used, after enlarging it, to store data that survives reboot.
A small partition of 2MB, large enough to hold an ext filesystem with
journalling is added to the img file.
The partition gets an ext2 filesystem as that is more efficient than
ext3. Since the second partition will maily hold browser cache data it
is assumed that it will be nostly read.

The first partition that contains the Webconverger runtime is made
larger to a size of 1022MB to allow for the greater amount of data that
is downloaded to update Webconverger through its auto-update feature.
This will allow this feature to work on systems running from a USB thumb
drive.
The extra space in the first partition is created by the Debian
live-build code. The size of 1022 allows for the whole img to be 1GB
The filesystem of the first partition used to be FAT16 but was changed
to FAT32 to enable larger partition sizes (in theory 999MB is FAT16
limit).

Code was added that was written by Matthijs Kooijman that creates a
second data partition when installing Webconverger on a HDD. This means
that in both situations (Webconverger running from a HDD and from a USB
thumb drive) the same behaviour of auto updates of Webconverger and the
availability of storage that survives a system boot.